+++ example/lib/in_app_webiew_example.screen.dart
@@ -123,7 +123,7 @@ class _InAppWebViewExampleScreenState extends State<InAppWebViewExampleScreen> {
                     key: webViewKey,
                     // contextMenu: contextMenu,
                     initialUrlRequest:
-                    URLRequest(url: Uri.parse("https://github.com/flutter")),
+                    URLRequest(url: Uri.parse("https://www.youtube.com/watch?v=7_v6oMtz7tA")),
                     // initialFile: "assets/index.html",
                     initialUserScripts: UnmodifiableListView<UserScript>([]),
                     initialOptions: options,
@@ -172,6 +172,8 @@ class _InAppWebViewExampleScreenState extends State<InAppWebViewExampleScreen> {
                         this.url = url.toString();
                         urlController.text = this.url;
                       });
+                      await Future.delayed(Duration(seconds: 2));
+                      await controller.closeAllMediaPresentations();
                     },
                     onLoadError: (controller, url, code, message) {
                       pullToRefreshController.endRefreshing();